rename: Slider 폴더 수정
Resolvers: #157
