+++ frontend/src/components/Slider/RangeMultiSlider_F/RangeMultiSlider_F.styled.ts
